spark/requirements.txt
flytekit == 1.10.0
flytekitplugins-spark == 1.10.0
flytekitplugins-envd == 1.10.0
